requirements.txt
absl-py==1.3.0
aioice==0.7.6
aiortc==1.3.2
altair==4.2.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
av==9.2.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.4
cv==1.0.0
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
distlib==0.3.6
dnspython==2.2.1
easyocr==1.6.2
entrypoints==0.4
executing==1.2.0
filelock==3.8.2
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.1.0
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
netifaces==0.11.0
networkx==2.8.8
ninja==1.11.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pydeck==0.8.0
pyee==9.0.4
Pygments==2.13.0
pylibsrtp==0.7.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
seaborn==0.12.1
semver==2.13.0
shapely==2.0.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.16.0
termcolor==2.1.1
tifffile==2022.10.10
toml==0.10.2
toolz==0.12.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
virtualenv==20.17.1
virtualenvwrapper-win==1.2.7
watchdog==2.2.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
